fix int type

+++ benchmarks/utils.cpp
@@ -130,7 +130,7 @@ int parseArgs(int argc, char ** argv,
     }
 
     auto max_key_string = getCmdOption(argv, argv + argc, "-k");
-    int max_key = LLONG_MAX;
+    long long int max_key = LLONG_MAX;
     if (!max_key_string.empty()) {
         max_key = std::stol(max_key_string);
     }
